+++ qoqo-qryd/tests/integration/tweezer_devices.rs
@@ -57,17 +57,37 @@ fn test_new() {
         assert_eq!(
             res.call_method0("seed")
                 .unwrap()
-                .extract::<usize>()
+                .extract::<Option<usize>>()
                 .unwrap(),
-            2
+            Some(2)
         );
         assert_eq!(
             res_mut
                 .call_method0("seed")
                 .unwrap()
-                .extract::<usize>()
+                .extract::<Option<usize>>()
                 .unwrap(),
-            2
+            Some(2)
+        );
+
+        let res_emp = device_type.call0().unwrap();
+        let res_mut_emp = device_type_mut.call0().unwrap();
+
+        assert_eq!(
+            res_emp
+                .call_method0("seed")
+                .unwrap()
+                .extract::<Option<usize>>()
+                .unwrap(),
+            None
+        );
+        assert_eq!(
+            res_mut_emp
+                .call_method0("seed")
+                .unwrap()
+                .extract::<Option<usize>>()
+                .unwrap(),
+            None
         );
     })
 }
@@ -821,15 +841,17 @@ fn test_two_qubit_edges() {
 fn test_from_api() {
     let mut sent_device = TweezerDevice::new(None, None, None);
     sent_device.add_layout("triangle").unwrap();
-    sent_device.set_tweezer_single_qubit_gate_time("PauliX", 0, 0.23, Some("triangle".to_string()));
+    sent_device.set_tweezer_single_qubit_gate_time(
+        "PhaseShiftState1",
+        0,
+        0.23,
+        Some("triangle".to_string()),
+    );
     sent_device.set_default_layout("triangle").unwrap();
-    // let sent_device_wrapper = TweezerDeviceWrapper {
-    //     internal: sent_device.clone(),
-    // };
     let mut received_device = TweezerDevice::new(None, None, None);
     received_device.add_layout("triangle").unwrap();
     received_device.set_tweezer_single_qubit_gate_time(
-        "PauliX",
+        "PhaseShiftState1",
         0,
         0.23,
         Some("triangle".to_string()),
@@ -867,7 +889,7 @@ fn test_from_api() {
         let device = device_type
             .call_method1(
                 "from_api",
-                (Option::<String>::None, Option::<String>::None, port),
+                (Option::<String>::None, Option::<String>::None, port, 42),
             )
             .unwrap();
 
@@ -880,6 +902,14 @@ fn test_from_api() {
                 .extract::<String>()
                 .unwrap(),
             "triangle"
+        );
+        assert_eq!(
+            device
+                .call_method0("seed")
+                .unwrap()
+                .extract::<usize>()
+                .unwrap(),
+            42
         );
 
         let returned_device_string = device