+++ roqoqo-qryd/src/api_backend.rs
@@ -14,7 +14,6 @@ use crate::api_devices::QRydAPIDevice;
 use bitvec::prelude::*;
 use num_complex::Complex64;
 use reqwest::blocking::Client;
-use reqwest::blocking::Response;
 use roqoqo::backends::RegisterResult;
 use roqoqo::measurements::ClassicalRegister;
 use roqoqo::operations::Define;
@@ -52,11 +51,15 @@ pub struct APIBackend {
     mock_port: Option<String>,
     /// Is develop version. Defaults to `false`.
     pub dev: bool,
+    /// API version.
+    api_version: String,
 }
 
 /// Local struct representing the body of the request message
 #[derive(Debug, serde::Serialize)]
 struct QRydRunData {
+    /// Format of the quantum program: qoqo
+    format: String,
     /// The QRyd WebAPI Backend used to execute operations and circuits.
     /// At the moment limited to the QRyd emulators
     /// ('qryd_emu_localcomp_square', 'qryd_emu_localcomp_triangle',
@@ -344,6 +347,7 @@ impl APIBackend {
     ///               been queried `timeout` times.
     /// * `mock_port` - Server port to be used for testing purposes.
     /// * `dev` - The boolean to set the dev option to.
+    /// * `api_version` - The version of the QRyd WebAPI to use. Defaults to "v3_0".
     ///
     pub fn new(
         device: QRydAPIDevice,
@@ -351,6 +355,7 @@ impl APIBackend {
         timeout: Option<usize>,
         mock_port: Option<String>,
         dev: Option<bool>,
+        api_version: Option<String>,
     ) -> Result<Self, RoqoqoBackendError> {
         if mock_port.is_some() {
             Ok(Self {
@@ -359,6 +364,7 @@ impl APIBackend {
                 timeout: timeout.unwrap_or(30),
                 mock_port,
                 dev: false,
+                api_version: api_version.unwrap_or("v3_0".to_string()),
             })
         } else {
             let access_token_internal: String = match access_token {
@@ -376,6 +382,7 @@ impl APIBackend {
                 timeout: timeout.unwrap_or(30),
                 mock_port,
                 dev: dev.unwrap_or(false),
+                api_version: api_version.unwrap_or("v3_0".to_string()),
             })
         }
     }
@@ -395,7 +402,7 @@ impl APIBackend {
     ///
     pub fn post_job(&self, quantumprogram: QuantumProgram) -> Result<String, RoqoqoBackendError> {
         // Prepare data that need to be passed to the WebAPI client
-        let seed_param: usize = self.device.seed(); // seed.unwrap_or(0);
+        let seed_param: Option<usize> = self.device.seed(); // seed.unwrap_or(0);
         let mut transform_pragma_repeated_measurement: bool = false;
 
         match &quantumprogram {
@@ -502,11 +509,12 @@ impl APIBackend {
         //     downconvert_roqoqo_version(quantumprogram)?;
         // dbg!(&serde_json::to_string(&quantumprogram).unwrap());
         let data = QRydRunData {
+            format: "qoqo".to_string(),
             backend: self.device.qrydbackend(),
             program: filtered_qp,
             dev: self.dev,
             fusion_max_qubits: 4,
-            seed_simulator: Some(seed_param),
+            seed_simulator: seed_param,
             seed_compiler: None,
             use_extended_set: true,
             use_reverse_traversal: true,
@@ -534,25 +542,40 @@ impl APIBackend {
         // Call WebAPI client
         // here: value for put() temporarily fixed.
         // needs to be derived dynamically based on the provided parameter 'qrydbackend'
-        let resp: Response;
-        if let Some(mock_port) = &self.mock_port {
-            resp = client
+        let resp = if let Some(mock_port) = &self.mock_port {
+            client
                 .post(format!("http://127.0.0.1:{}", mock_port))
                 .json(&data)
                 .send()
                 .map_err(|e| RoqoqoBackendError::NetworkError {
                     msg: format!("{:?}", e),
-                })?;
+                })?
+        } else if self.dev {
+            client
+                .post(format!(
+                    "https://api.qryddemo.itp3.uni-stuttgart.de/{}/jobs",
+                    self.api_version
+                ))
+                .header("X-API-KEY", self.access_token.clone())
+                .header("X-DEV", "?1")
+                .json(&data)
+                .send()
+                .map_err(|e| RoqoqoBackendError::NetworkError {
+                    msg: format!("{:?}", e),
+                })?
         } else {
-            resp = client
-                .post("https://api.qryddemo.itp3.uni-stuttgart.de/v3_0/jobs")
+            client
+                .post(format!(
+                    "https://api.qryddemo.itp3.uni-stuttgart.de/{}/jobs",
+                    self.api_version
+                ))
                 .header("X-API-KEY", self.access_token.clone())
                 .json(&data)
                 .send()
                 .map_err(|e| RoqoqoBackendError::NetworkError {
                     msg: format!("{:?}", e),
-                })?;
-        }
+                })?
+        };
 
         let status_code = resp.status();
         if status_code != reqwest::StatusCode::CREATED {
@@ -625,14 +648,26 @@ impl APIBackend {
         let url_string: String = job_location + "/status";
 
         // Call WebAPI client
-        let resp = client
-            .get(url_string)
-            .header("X-API-KEY", self.access_token.clone())
-            // .json(&data)
-            .send()
-            .map_err(|e| RoqoqoBackendError::NetworkError {
-                msg: format!("{:?}", e),
-            })?;
+        let resp = if self.dev {
+            client
+                .get(url_string)
+                .header("X-API-KEY", self.access_token.clone())
+                .header("X-DEV", "?1")
+                // .json(&data)
+                .send()
+                .map_err(|e| RoqoqoBackendError::NetworkError {
+                    msg: format!("{:?}", e),
+                })?
+        } else {
+            client
+                .get(url_string)
+                .header("X-API-KEY", self.access_token.clone())
+                // .json(&data)
+                .send()
+                .map_err(|e| RoqoqoBackendError::NetworkError {
+                    msg: format!("{:?}", e),
+                })?
+        };
 
         let status_code = resp.status();
         if status_code != reqwest::StatusCode::OK {
@@ -698,14 +733,26 @@ impl APIBackend {
         let url_string: String = job_location + "/result";
 
         // Call WebAPI client
-        let resp = client
-            .get(url_string)
-            .header("X-API-KEY", self.access_token.clone())
-            // .json(&data)
-            .send()
-            .map_err(|e| RoqoqoBackendError::NetworkError {
-                msg: format!("{:?}", e),
-            })?;
+        let resp = if self.dev {
+            client
+                .get(url_string)
+                .header("X-API-KEY", self.access_token.clone())
+                .header("X-DEV", "?1")
+                // .json(&data)
+                .send()
+                .map_err(|e| RoqoqoBackendError::NetworkError {
+                    msg: format!("{:?}", e),
+                })?
+        } else {
+            client
+                .get(url_string)
+                .header("X-API-KEY", self.access_token.clone())
+                // .json(&data)
+                .send()
+                .map_err(|e| RoqoqoBackendError::NetworkError {
+                    msg: format!("{:?}", e),
+                })?
+        };
 
         let status_code = resp.status();
         if status_code != reqwest::StatusCode::OK {
@@ -763,13 +810,24 @@ impl APIBackend {
                 })?
         };
         // Call WebAPI client
-        let resp = client
-            .delete(job_location)
-            .header("X-API-KEY", self.access_token.clone())
-            .send()
-            .map_err(|e| RoqoqoBackendError::NetworkError {
-                msg: format!("{:?}", e),
-            })?;
+        let resp = if self.dev {
+            client
+                .delete(job_location)
+                .header("X-API-KEY", self.access_token.clone())
+                .header("X-DEV", "?1")
+                .send()
+                .map_err(|e| RoqoqoBackendError::NetworkError {
+                    msg: format!("{:?}", e),
+                })?
+        } else {
+            client
+                .delete(job_location)
+                .header("X-API-KEY", self.access_token.clone())
+                .send()
+                .map_err(|e| RoqoqoBackendError::NetworkError {
+                    msg: format!("{:?}", e),
+                })?
+        };
 
         let status_code = resp.status();
         if status_code != reqwest::StatusCode::OK {
@@ -966,11 +1024,19 @@ mod test {
     #[test]
     fn debug_and_clone() {
         let device: QRydAPIDevice = QrydEmuSquareDevice::new(None, None, None).into();
-        let backend =
-            APIBackend::new(device.clone(), Some("".to_string()), Some(2), None, None).unwrap();
+        let backend = APIBackend::new(
+            device.clone(),
+            Some("".to_string()),
+            Some(2),
+            None,
+            None,
+            None,
+        )
+        .unwrap();
         let a = format!("{:?}", backend);
         assert!(a.contains("QrydEmuSquareDevice"));
-        let backend2 = APIBackend::new(device, Some("a".to_string()), Some(2), None, None).unwrap();
+        let backend2 =
+            APIBackend::new(device, Some("a".to_string()), Some(2), None, None, None).unwrap();
         assert_eq!(backend.clone(), backend);
         assert_ne!(backend, backend2);
     }
@@ -990,6 +1056,7 @@ mod test {
         // let program: roqoqo_1_0::QuantumProgram = downconvert_roqoqo_version(program).unwrap();
 
         let test = QRydRunData {
+            format: "qoqo".to_string(),
             backend: "qryd_emu_cloudcomp_square".to_string(),
             program,
             dev: false,
@@ -1002,7 +1069,7 @@ mod test {
             extended_set_weight: 0.5,
             reverse_traversal_iterations: 2,
         };
-        assert_eq!(format!("{:?}", test), "QRydRunData { backend: \"qryd_emu_cloudcomp_square\", dev: false, fusion_max_qubits: 4, seed_simulator: None, seed_compiler: None, use_extended_set: true, use_reverse_traversal: true, reverse_traversal_iterations: 2, extended_set_size: 5, extended_set_weight: 0.5, program: ClassicalRegister { measurement: ClassicalRegister { constant_circuit: None, circuits: [Circuit { definitions: [], operations: [], _roqoqo_version: RoqoqoVersion }] }, input_parameter_names: [\"test\"] } }");
+        assert_eq!(format!("{:?}", test), "QRydRunData { format: \"qoqo\", backend: \"qryd_emu_cloudcomp_square\", dev: false, fusion_max_qubits: 4, seed_simulator: None, seed_compiler: None, use_extended_set: true, use_reverse_traversal: true, reverse_traversal_iterations: 2, extended_set_size: 5, extended_set_weight: 0.5, program: ClassicalRegister { measurement: ClassicalRegister { constant_circuit: None, circuits: [Circuit { definitions: [], operations: [], _roqoqo_version: RoqoqoVersion }] }, input_parameter_names: [\"test\"] } }");
     }
 
     /// Test Debug of QRydJobResult
@@ -1088,7 +1155,8 @@ mod test {
         let number_qubits = 6;
         let device = QrydEmuSquareDevice::new(Some(2), None, None);
         let qryd_device: QRydAPIDevice = QRydAPIDevice::from(&device);
-        let api_backend_new = APIBackend::new(qryd_device, None, None, Some(port), None).unwrap();
+        let api_backend_new =
+            APIBackend::new(qryd_device, None, None, Some(port), None, None).unwrap();
         let mut circuit = Circuit::new();
         circuit += operations::DefinitionBit::new("ro".to_string(), number_qubits, true);
         circuit += operations::RotateX::new(0, std::f64::consts::PI.into());
@@ -1173,7 +1241,8 @@ mod test {
         let number_qubits = 6;
         let device = QrydEmuSquareDevice::new(Some(2), None, None);
         let qryd_device: QRydAPIDevice = QRydAPIDevice::from(&device);
-        let api_backend_new = APIBackend::new(qryd_device, None, None, Some(port), None).unwrap();
+        let api_backend_new =
+            APIBackend::new(qryd_device, None, None, Some(port), None, None).unwrap();
         let mut circuit = Circuit::new();
         circuit += operations::DefinitionBit::new("ro".to_string(), number_qubits, true);
         circuit += operations::RotateX::new(0, std::f64::consts::PI.into());
@@ -1240,7 +1309,8 @@ mod test {
             .collect::<String>();
         let device = QrydEmuSquareDevice::new(Some(1), None, None);
         let qryd_device: QRydAPIDevice = QRydAPIDevice::from(&device);
-        let api_backend_new = APIBackend::new(qryd_device, None, None, Some(port), None).unwrap();
+        let api_backend_new =
+            APIBackend::new(qryd_device, None, None, Some(port), None, None).unwrap();
 
         let mut input_circuit = Circuit::new();
         input_circuit += operations::DefinitionBit::new("ro".to_string(), 3, true);
@@ -1292,6 +1362,7 @@ mod test {
             input_parameter_names: vec![],
         };
         let data = QRydRunData {
+            format: "qoqo".to_string(),
             backend: device.qrydbackend(),
             program: output_program,
             dev: false,