+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 Tracks qoqo-qryd changes after 0.5
 
+# 0.11.4
+
+* Added `api_version`, `seed`, `dev` parameters to `TweezerDevice.from_api()`
+* Fixed `TweezerDevice` support for `APIBackend`
+* Modified `TweezerDevice` seed parameter to default to None, instead of 0
+* Added `api_version` parameter to `APIBackend`
+
 # 0.11.3
 
 * Modified `TweezerDevice.from_json()` and `TweezerMutableDevice.set_default_layout()` to automatically switch the layout of the device if a default one was set